Change type of shadows from array to LayerDrawable
Summary:
Before we were using lists which meant `CompositeBackgroundDrawable` had a `LayerDrawable` of variable length. 

This meant that our layers were not consistent and made it hard to optimize `CompositeBackgroundDrawable` to not create a new instance every time we set a new Layer.  

With this change `CompositeBackgroundDrawable` has a consistent amount of layers which will allow us to more easily mutate it and land the optimization present on D63287222

Differential Revision: D63798842